Created contexts Chat and Accounts. Backend functionality works.
