+++ stix_server/lib/stix_server/schemas/message.ex
@@ -1,10 +1,6 @@
 defmodule StixServer.Schemas.Message do
   use Ecto.Schema
   import Ecto.Changeset
-  import Ecto.Query, only: [from: 2]
-  
-  alias StixServer.Schemas.Message
-  alias StixServer.Repo
 
   @derive {Jason.Encoder, only: [:id, :message_body, :dialogue_id, :sender_id, :receiver_id]}
   schema "messages" do
@@ -21,16 +17,5 @@ defmodule StixServer.Schemas.Message do
     message
     |> cast(attrs, [:message_body, :dialogue_id, :sender_id, :receiver_id])
     |> validate_required([:message_body, :dialogue_id, :sender_id, :receiver_id])
-  end
-
-  def create_message(params) do
-    Message.changeset(%Message{}, params) |> Repo.insert
-  end
-
-  def get_messages(dialogue_id) do
-    (from m in Message, 
-      where: (m.dialogue_id == ^dialogue_id),
-      select: m) 
-      |> Repo.all()
   end
 end
